refactor(economy): keep the engine out of the renderer's import graph
Closes #610
