+++ CLAUDE.md
@@ -129,7 +129,7 @@ Tests share their isolation in `tests/conftest.py` and their scaffolding in `tes
 - Casino-system P&L lives in the `casino_ledger` row in `economy.db`; `/casino` reads the ledger, `/pocat` reads the bot's wallet. Do not reuse the bot's `user_wallet` as the house ledger. Player wallet and casino ledger commit in one atomic transaction, and Blackjack enters through `apply_blackjack_settlement` — the wrapper that exists to keep the five-card-21 bonus off the ledger — rather than `apply_round_settlement` directly.
 - Ephemeral vs public boundary: shared social / market / settlement events are public embeds with scheduled cleanup; personal state and validation / permission failures are ephemeral.
 - Exactly one faucet pays an action reward: `cli.py`'s cooldown-gated per-message reward. Other cogs must not invent a second one; the system-funded mints that settle a loan or a game are not rewards.
-- Anti-inflation levers: `transfer()` burns `TRANSFER_TAX_BPS` of every `/give` as a permanent sink and casino wagers cap at `MAX_SINGLE_BET` at the shared chokepoints (both constants in `typings/economy.py`); the VIP casino multiplier is 1.2x via `apply_vip_blackjack_bonus` (`_VIP_WIN_MULTIPLIER_NUM`/`_DEN` in `services/economy/database.py`). Re-measure before changing.
+- Anti-inflation levers: `transfer()` burns `TRANSFER_TAX_BPS` of every `/give` as a permanent sink and casino wagers cap at `MAX_SINGLE_BET` at the shared chokepoints (both constants in `typings/economy.py`); the VIP casino multiplier is 1.2x via `apply_vip_blackjack_bonus` (`_VIP_WIN_MULTIPLIER_NUM`/`_DEN`), which lives in `typings/economy.py` beside them along with the two loan-rate converters, so nothing that only formats or settles a number has to import the ledger's engine to reach one (#610). Re-measure before changing.
 
 ## Games
